requirements.txt
requests
selenium
beautifulsoup4
win11toast
python-dotenv
schedule
pystray
Pillow
